+++ Gems/ScriptCanvasTesting/Code/Tests/ScriptCanvas_NodeGenerics.cpp
@@ -110,7 +110,7 @@ namespace ScriptCanvas
     AZ_INLINE AZ::Vector3 NormalizeWithDefault(const AZ::Vector3& source, const Data::NumberType tolerance, [[maybe_unused]] const Data::BooleanType fakeValueForTestingDefault)
     {
         AZ_TracePrintf("SC", "The fake value for testing default is %s\n", fakeValueForTestingDefault ? "True" : "False");
-        return source.GetNormalizedSafe(tolerance);
+        return source.GetNormalizedSafe(static_cast<float>(tolerance));
     }
 
     void NormalizeWithDefaultInputOverrides(Node& node) { SetDefaultValuesByIndex< 1, 2 >::_(node, 3.3, true); }
